Remove the login option on report.html
This worked great on a fake instance of the login
menu but doesn't work when we're dealing with cross-origin
login pages.

Simplified the logic to show a nag (dialog) if the user
is not logged in. This is not great as we've lost the
current report but at least it works.